+++ game/assets/objects/advanced_rigid_body.gd
@@ -39,6 +39,14 @@ var _original_albedo_arr := PoolColorArray()
 # A flag which determines if we can safely read _original_albedo_arr.
 var _original_albedo_saved := false
 
+# The list of transforms corresponding to each of the scene's collision shapes.
+# NOTE: This system assumes that the node structure does not change through the
+# course of play.
+var _original_transform_arr := []
+
+# A flag which determines if we can safely read _original_transform_arr.
+var _original_transform_saved := false
+
 
 ## Get the list of [CollisionShape] nodes that this rigid body uses.
 ## TODO: Make typed in 4.x
@@ -87,12 +95,12 @@ func get_materials() -> Array:
 ## Get the custom albedo colour applied to this body by the player.
 func get_user_albedo() -> Color:
 	var material_arr := get_materials()
-	if material_arr.size() < 1:
+	if material_arr.empty():
 		return Color.white
 	
 	var first_material: SpatialMaterial = material_arr[0]
 	var initial_albedo_arr := _get_original_albedo_arr()
-	if initial_albedo_arr.size() < 1:
+	if initial_albedo_arr.empty():
 		return Color.white
 	
 	# We don't want to get just the current albedo, instead we want to get the
@@ -131,6 +139,56 @@ func set_user_albedo(new_albedo: Color) -> void:
 		material.albedo_color = initial_albedo * new_albedo
 
 
+## Get the custom scale applied to this body.
+func get_user_scale() -> Vector3:
+	var collision_shape_arr := get_collision_shapes()
+	if collision_shape_arr.empty():
+		return Vector3.ONE
+	
+	var first_collision_shape: CollisionShape = collision_shape_arr[0]
+	var initial_transform_arr := _get_original_transform_arr()
+	if initial_transform_arr.empty():
+		return Vector3.ONE
+	
+	# The custom scale should have been set for all collision shapes, so it
+	# doesn't matter which collision shape we use, so we'll use the first one.
+	# NOTE: We don't want to return just the current scale, we need to return
+	# the scale RELATIVE to the initial scale.
+	var first_initial_transform: Transform = initial_transform_arr[0]
+	var first_current_transform: Transform = first_collision_shape.transform
+	
+	var initial_scale := first_initial_transform.basis.get_scale()
+	var current_scale := first_current_transform.basis.get_scale()
+	
+	var x := 1.0 if is_zero_approx(initial_scale.x) else \
+			current_scale.x / initial_scale.x
+	var y := 1.0 if is_zero_approx(initial_scale.y) else \
+			current_scale.y / initial_scale.y
+	var z := 1.0 if is_zero_approx(initial_scale.z) else \
+			current_scale.z / initial_scale.z
+	
+	return Vector3(x, y, z)
+
+
+## Set the custom scale applied to this body.
+func set_user_scale(new_scale: Vector3) -> void:
+	var collision_shape_arr := get_collision_shapes()
+	var initial_transform_arr := _get_original_transform_arr()
+	
+	if collision_shape_arr.size() != initial_transform_arr.size():
+		push_error("Number of collision shapes '%d' does not match number of initial transforms '%d'" % [
+				collision_shape_arr.size(), initial_transform_arr.size()])
+		return
+	
+	for index in range(collision_shape_arr.size()):
+		var collision_shape: CollisionShape = collision_shape_arr[index]
+		var initial_transform: Transform = initial_transform_arr[index]
+		
+		collision_shape.transform = initial_transform.scaled(new_scale)
+	
+	reset_physics_interpolation()
+
+
 # Get the original albedo colours assigned to the materials that belong to this
 # rigid body. The first time this function is called, the results are saved in
 # the event that the albedo colours are changed.
@@ -144,3 +202,19 @@ func _get_original_albedo_arr() -> PoolColorArray:
 		_original_albedo_saved = true
 	
 	return _original_albedo_arr
+
+
+# Get the original transforms assigned to each of this body's collision shapes.
+# The first time this function is called, the results are saved in the event
+# that the transforms are changed.
+# TODO: Make typed in 4.x
+func _get_original_transform_arr() -> Array:
+	if not _original_transform_saved:
+		_original_albedo_arr = []
+		for element in get_collision_shapes():
+			var collision_shape: CollisionShape = element
+			_original_transform_arr.push_back(collision_shape.transform)
+		
+		_original_transform_saved = true
+	
+	return _original_transform_arr